Added nested samping package, moved some code to a nested ABC
