+++ autofit/optimize/non_linear/__init__.py
@@ -1,6 +1,6 @@
 from .downhill_simplex import DownhillSimplex
 from .grid_search import GridSearch
-from .dynesty import Dynesty
+from .nested_sampling.dynesty import Dynesty
 from .emcee import Emcee
-from .multi_nest import MultiNest
+from .nested_sampling.multi_nest import MultiNest
 from .non_linear import NonLinearOptimizer
