+++ autofit/optimize/non_linear/output.py
@@ -376,53 +376,3 @@ class AbstractOutput:
 
 class MCMCOutput(AbstractOutput):
     pass
-
-
-class NestedSamplingOutput(AbstractOutput):
-    def weight_from_sample_index(self, sample_index):
-        raise NotImplementedError()
-
-    @property
-    def evidence(self):
-        raise NotImplementedError()
-
-    def output_pdf_plots(self):
-
-        import getdist.plots
-        import matplotlib
-
-        backend = conf.instance.visualize_general.get("general", "backend", str)
-        if not backend in "default":
-            matplotlib.use(backend)
-        import matplotlib.pyplot as plt
-
-        pdf_plot = getdist.plots.GetDistPlotter()
-
-        plot_pdf_1d_params = conf.instance.visualize_plots.get("pdf", "1d_params", bool)
-
-        if plot_pdf_1d_params:
-
-            for param_name in self.model.param_names:
-                pdf_plot.plot_1d(roots=self.pdf, param=param_name)
-                pdf_plot.export(
-                    fname="{}/pdf_{}_1D.png".format(self.paths.pdf_path, param_name)
-                )
-
-        plt.close()
-
-        plot_pdf_triangle = conf.instance.visualize_plots.get("pdf", "triangle", bool)
-
-        if plot_pdf_triangle:
-
-            try:
-                pdf_plot.triangle_plot(roots=self.pdf)
-                pdf_plot.export(fname="{}/pdf_triangle.png".format(self.paths.pdf_path))
-            except Exception as e:
-                print(type(e))
-                print(
-                    "The PDF triangle of this non-linear search could not be plotted. This is most likely due to a "
-                    "lack of smoothness in the sampling of parameter space. Sampler further by decreasing the "
-                    "parameter evidence_tolerance."
-                )
-
-        plt.close()
