+++ test_autofit/unit/optimize/files/dynesty/config/visualize/general.ini
@@ -1,0 +1,3 @@
+[general]
+backend = TKAgg
+visualize_interval = -1
