+++ test_autofit/unit/optimize/files/emcee/config/visualize/plots.ini
@@ -1,0 +1,3 @@
+[pdf]
+triangle = False
+1d_params = False
